Trajectory plots in ggplot2

These functions existed at a time when openair::trajPlot() and openair::trajLevel() were written in lattice. Now they are written in ggplot2, these functions have been deprecated and now pass all of their arguments to their corresponding openair functions. They will be removed in a future version of openairmaps.

Usage

trajLevelMapStatic(...)

trajMapStatic(...)

Arguments

...

Arguments passed to either openair::trajPlot() or openair::trajLevel().