Add polar markers to leaflet map

This function is similar (but not identical to) the leaflet::addMarkers() and leaflet::addCircleMarkers() functions in leaflet, which allows users to add openair directional analysis plots to any leaflet map and have more control over groups and layerIds than in "all-in-one" functions like polarMap().

Usage

addPolarMarkers(
  map,
  data,
  pollutant,
  fun = openair::polarPlot,
  lng = NULL,
  lat = NULL,
  layerId = NULL,
  group = NULL,
  popup = NULL,
  label = NULL,
  key = FALSE,
  d.icon = 200,
  d.fig = 3.5,
  ...
)

addPolarDiffMarkers(
  map,
  before,
  after,
  pollutant,
  lng = NULL,
  lat = NULL,
  layerId = NULL,
  group = NULL,
  popup = NULL,
  label = NULL,
  key = FALSE,
  d.icon = 200,
  d.fig = 3.5,
  ...
)

Arguments

map

a map widget object created from leaflet()

data

A data frame. The data frame must contain the data to plot your choice of openair directional analysis plot, which includes wind speed (ws), wind direction (wd), and the column representing the concentration of a pollutant. In addition, data must include a decimal latitude and longitude.

pollutant

The name of the pollutant to be plot. Note that, if fun = openair::windRose, you must set pollutant = "ws".

fun

An openair directional analysis plotting function. Supported functions include openair::polarPlot() (the default), openair::polarAnnulus(), openair::polarFreq(), openair::percentileRose(), openair::pollutionRose() and openair::windRose(). For openair::polarDiff(), use addPolarDiffMarkers().

lng

The decimal longitude.

lat

The decimal latitude.

layerId

the layer id

group

the name of the group the newly created layers should belong to (for clearGroup and addLayersControl purposes). Human-friendly group names are permitted--they need not be short, identifier-style names. Any number of layers and even different types of layers (e.g. markers and polygons) can share the same group name.

popup

A column of data to be used as a popup.

label

A column of data to be used as a label.

key

Should a key for each marker be drawn? Default is FALSE.

d.icon

The diameter of the plot on the map in pixels. This will affect the size of the individual polar markers. Alternatively, a vector in the form c(width, height) can be provided if a non-circular marker is desired.

d.fig

The diameter of the plots to be produced using openair in inches. This will affect the resolution of the markers on the map. Alternatively, a vector in the form c(width, height) can be provided if a non-circular marker is desired.

...

Other arguments for the plotting function (e.g. period for openair::polarAnnulus()).

before

A data frame that represents the "before" case. See polarPlot() for details of different input requirements.

after

A data frame that represents the "after" case. See polarPlot() for details of different input requirements.

Value

A leaflet object.

Functions

addPolarMarkers(): Add any one-table polar marker (e.g., openair::polarPlot())

addPolarDiffMarkers(): Add the two-table openair::polarDiff() marker.

Examples

if (FALSE) {
library(leaflet)
library(openair)

# different types of polar plot on one map
leaflet() %>%
  addTiles() %>%
  addPolarMarkers(
    data = polar_data,
    pollutant = "ws",
    fun = windRose,
    group = "Wind Rose"
  ) %>%
  addPolarMarkers(
    data = polar_data,
    pollutant = "nox",
    fun = polarPlot,
    group = "Polar Plot"
  ) %>%
  addLayersControl(
    baseGroups = c("Wind Rose", "Polar Plot")
  )

# use of polar diff
leaflet() %>%
  addTiles() %>%
  addPolarDiffMarkers(
    before = polar_data,
    after = dplyr::mutate(polar_data, nox = jitter(nox, 5)),
    pollutant = "nox"
  )
}